+++ PATNTupler/interface/JetStruct.hh
@@ -9,20 +9,20 @@ namespace ran{
   struct JetStruct: public TObject{
     JetStruct(); ///< default constructor
     virtual ~JetStruct(){} ///< destructor
-    double et;///< et
-    double pt;/// < pt
-    double mass;/// < mass
-    double eta;///< eta 
-    double phi;/// < phi
-    double jetProbabilityBJetTags;///< jetProbabilityBJetTags b tag discriminator
-    double jetBProbabilityBJetTags;///< jetBProbabilityBJetTags b tag discriminator
-    double trackCountingHighEffBJetTags;///< trackCountingHighEffBJetTags b tag discriminator
-    double trackCountingHighPurBJetTags;///< trackCountingHighPurBJetTags b tag discriminator
-    double simpleSecondaryVertexHighEffBJetTags;///< simpleSecondaryVertexHighEffBJetTags b tag discriminator
-    double simpleSecondaryVertexHighPurBJetTags;///< combinedSecondaryVertexBJetTags b tag discriminator
-    double combinedInclusiveSecondaryVertexBJetTags;///< combinedInclusiveSecondaryVertexBJetTags b tag discriminator
-    double jecFactor_unCorrected;///<factor applied to jet pt to get uncorrect jet pt
-    double userFloat_pileupJetId_fullDiscriminant;///< userFloat pileup jet ID
+    float et;///< et
+    float pt;/// < pt
+    float mass;/// < mass
+    float eta;///< eta 
+    float phi;/// < phi
+    float jetProbabilityBJetTags;///< jetProbabilityBJetTags b tag discriminator
+    float jetBProbabilityBJetTags;///< jetBProbabilityBJetTags b tag discriminator
+    float trackCountingHighEffBJetTags;///< trackCountingHighEffBJetTags b tag discriminator
+    float trackCountingHighPurBJetTags;///< trackCountingHighPurBJetTags b tag discriminator
+    float simpleSecondaryVertexHighEffBJetTags;///< simpleSecondaryVertexHighEffBJetTags b tag discriminator
+    float simpleSecondaryVertexHighPurBJetTags;///< combinedSecondaryVertexBJetTags b tag discriminator
+    float combinedInclusiveSecondaryVertexBJetTags;///< combinedInclusiveSecondaryVertexBJetTags b tag discriminator
+    float jecFactor_unCorrected;///<factor applied to jet pt to get uncorrect jet pt
+    float userFloat_pileupJetId_fullDiscriminant;///< userFloat pileup jet ID
     int partonFlavour;///<  MC parton flavour (sensible values for MC only obviously!) 
     ClassDef(JetStruct,1)//Struct class definition
   };
